move mosaic cog process to georeference app
