+++ ohmg/georeference/management/commands/mosaic.py
@@ -1,6 +1,9 @@
 from argparse import Namespace
 from django.core.management.base import BaseCommand
+
 from ohmg.core.models import LayerSet
+from ohmg.georeference.mosaicker import Mosaicker
+from ohmg.georeference.tasks import create_mosaic_cog
 
 
 class Command(BaseCommand):
@@ -10,13 +13,13 @@ class Command(BaseCommand):
         parser.add_argument(
             "operation",
             choices=[
-                "inspect",
+                "generate-cog",
             ],
             help="the operation to perform",
         )
         parser.add_argument(
             "-i",
-            "--identifier",
+            "--mapid",
             help="the identifier of map that holds this layerset",
         )
         parser.add_argument(
@@ -24,6 +27,14 @@ class Command(BaseCommand):
             "--category",
             default="main-content",
             help="category of the layerset to work with",
+        )
+        parser.add_argument(
+            "--pk",
+            help="a pk can also be used to identify the layerset",
+        )
+        parser.add_argument(
+            "--background",
+            action="store_true",
         )
 
     def handle(self, *args, **options):
@@ -33,8 +44,11 @@ class Command(BaseCommand):
             ls = LayerSet.objects.get(pk=options.pk)
         else:
             ls = LayerSet.objects.get(
-                map__identifier=options.identifier, category__slug=options.category
+                map__identifier=options.mapid, category__slug=options.category
             )
 
-        if options.operation == "inspect":
-            print(ls.multimask_extent)
+        if options.operation == "generate-cog":
+            if options.background:
+                create_mosaic_cog.delay(ls.pk)
+            else:
+                Mosaicker().generate_cog(ls)
